refactor: make extension over DbContext
